fix payments and loading books
