+++ src/api/requests/token-tracker.ts
@@ -1,5 +1,5 @@
 import { api } from '@/api'
-import { Payment, PageOrder, NftPayment } from '@/types'
+import { Payment, PageOrder } from '@/types'
 import { config } from '@/config'
 
 export function getPayments(opts: {
@@ -10,28 +10,6 @@ export function getPayments(opts: {
   pageOrder?: PageOrder
 }) {
   return api.get<Payment[]>('/integrations/tracker/payments', {
-    page: {
-      limit: opts.pageLimit || config.DEFAULT_PAGE_LIMIT,
-      order: opts.pageOrder || 'desc',
-    },
-    filter: {
-      ...(opts?.bookIds?.length ? { book_id: opts.bookIds.join(',') } : {}),
-      ...(opts?.ids?.length ? { id: opts.ids.join(',') } : {}),
-      ...(opts?.tokenAddresses?.length
-        ? { token_address: opts.tokenAddresses.join(',') }
-        : {}),
-    },
-  })
-}
-
-export function getNftPayments(opts: {
-  bookIds?: (number | string)[]
-  ids?: (number | string)[]
-  tokenAddresses?: string[]
-  pageLimit?: number
-  pageOrder?: PageOrder
-}) {
-  return api.get<NftPayment[]>('/integrations/tracker/payments/nft', {
     page: {
       limit: opts.pageLimit || config.DEFAULT_PAGE_LIMIT,
       order: opts.pageOrder || 'desc',
