+++ src/enums/index.ts
@@ -8,3 +8,4 @@ export * from './fields.enum'
 export * from './promocodes.enum'
 export * from './networks.enum'
 export * from './currencies.enum'
+export * from './tokens.enum'
