+++ src/enums/tokens.enum.ts
@@ -1,0 +1,6 @@
+export enum TOKEN_TYPES {
+  native = 'native',
+  erc20 = 'erc20',
+  voucher = 'voucher',
+  nft = 'nft',
+}
