+++ src/pages/nft-item-page/index.ts
@@ -1,5 +1,3 @@
 export { default as NftDetails } from './NftDetails.vue'
 export { default as SaleHistory } from './SaleHistory.vue'
 export { default as SaleHistoryItem } from './SaleHistoryItem.vue'
-export { default as SaleHistoryTokens } from './SaleHistoryTokens.vue'
-export { default as SaleHistoryNfts } from './SaleHistoryNfts.vue'
